+++ jobmon_core/pyproject.toml
@@ -24,6 +24,8 @@ dependencies = [
     'grpcio<=1.64.1',
     'docutils',
     'sphinx',
+    'opentelemetry-api',
+    'opentelemetry-sdk',
 ]
 dynamic = ["version"]
